requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-pinecone
pinecone-client
pypdf
langchain_community